refactor(midnight-js): read the hex fixtures through one shared helper
readHexFixture was copied byte-identically into seven suites, one of which had already drifted into inlining the cross-package fixture path at its call site.
