+++ packages/protocol/src/test/era-load-era.test.ts
@@ -13,8 +13,6 @@
  * limitations under the License.
  */
 
-import { readFileSync } from 'node:fs';
-import { resolve } from 'node:path';
 
 import * as ocrt3 from '@midnight-ntwrk/onchain-runtime-v3';
 import { describe, expect, it } from 'vitest';
@@ -22,17 +20,9 @@ import { describe, expect, it } from 'vitest';
 import { PROTOCOL_ERROR_CODES, UnknownLedgerVersionError } from '../errors';
 import { extractEncodedStateValue } from '../lib/engine/envelope';
 import { loadLedgerEra } from '../lib/era/load-era';
+import { readHexFixture } from './fixtures';
 
-const FIXTURES_DIR = resolve(__dirname, '../../../../testkit-js/testkit-js/src/fixtures/hf');
 
-const readHexFixture = (name: string): Uint8Array => {
-  const text = readFileSync(resolve(FIXTURES_DIR, name), 'utf8').trim();
-  const bytes = Uint8Array.from(Buffer.from(text, 'hex'));
-  if (bytes.length * 2 !== text.length) {
-    throw new Error(`fixture ${name} is not valid hex in full: ${text.length} chars decoded to ${bytes.length} bytes`);
-  }
-  return bytes;
-};
 
 // The golden pair: the same contract before and after the migration. Reading
 // each through its own era is what makes the two arms comparable at all.
